+++ src/mbedtls/session.cpp
@@ -96,7 +96,7 @@ void MbedTlsSession::performHttpsPrologue(std::string const& hostname, bool veri
             std::array<char, 512> buffer;
             auto const n = mbedtls_x509_crt_verify_info(buffer.data(), buffer.size(), "", flags);
             if (n > 0) {
-                buffer[n - 1] = '\0';  // get rid of the newline
+                buffer[static_cast<size_t>(n - 1)] = '\0';  // get rid of the newline
             }
             spdlog::error("Certificate verification failed: {}", buffer.data());
             throw OhcException{"certification verification failed"};
